refactors

+++ packages/@livestore/web/src/storage/web-worker/index.ts
@@ -1,8 +1,19 @@
-import type { StorageDatabase } from '@livestore/common'
+import type { ResetMode, StorageDatabase } from '@livestore/common'
 import { makeSchemaHash } from '@livestore/common/schema'
 import { casesHandled, omit } from '@livestore/utils'
-import type { Duration } from '@livestore/utils/effect'
-import { BrowserWorker, Deferred, Effect, Exit, FiberId, Queue, Scope, WebLock, Worker } from '@livestore/utils/effect'
+import {
+  BrowserWorker,
+  Chunk,
+  Deferred,
+  Effect,
+  Exit,
+  FiberId,
+  pipe,
+  Queue,
+  Scope,
+  WebLock,
+  Worker,
+} from '@livestore/utils/effect'
 
 import { IDB } from '../utils/idb.js'
 import type { StorageInit } from '../utils/types.js'
@@ -15,8 +26,8 @@ import {
 } from './common.js'
 import * as WorkerSchema from './schema.js'
 
+/** Specifies where to persist data for this storage */
 export type StorageOptionsWeb = {
-  /** Specifies where to persist data for this storage */
   worker: Worker | (new (options?: { name: string }) => Worker)
 } & WorkerSchema.StorageType
 
@@ -31,35 +42,35 @@ export const createStorage =
   ({ otel: {}, schema }) => {
     const manualScope = Effect.runSync(Scope.make())
 
-    return Effect.gen(function* ($) {
-      const executionBacklogQueue = yield* $(
-        Queue.unbounded<WorkerSchema.ExecutionBacklogItem>(),
+    return Effect.gen(function* () {
+      const executionBacklogQueue = yield* Queue.unbounded<WorkerSchema.ExecutionBacklogItem>().pipe(
         Effect.acquireRelease(Queue.shutdown),
       )
 
-      const lockDeferred = yield* $(Deferred.make<void>())
+      const lockDeferred = yield* pipe(Deferred.make<void>())
 
       const worker =
         options.worker instanceof globalThis.Worker ? options.worker : new options.worker({ name: 'livestore-worker' })
 
       const storageOptions = omit(options, ['worker'])
 
-      const workerDeferred = yield* $(
+      const workerDeferred = yield* pipe(
         Worker.makePoolSerialized<WorkerSchema.Request>({
           size: 1,
           permits: 10,
-          initialMessage: () => new WorkerSchema.InitialMessage({ storage: storageOptions }),
+          initialMessage: () => new WorkerSchema.InitialMessage({ storageOptions }),
         }).pipe(Effect.provide(BrowserWorker.layer(() => worker)), Effect.toForkedDeferred),
       )
 
       const runInWorker = <TReq extends WorkerSchema.Request>(req: TReq) =>
         Effect.andThen(Deferred.await(workerDeferred), (worker) => worker.executeEffect(req))
 
-      yield* $(
-        Deferred.await(workerDeferred),
+      yield* Deferred.await(workerDeferred).pipe(
         Effect.andThen(() =>
-          queueTakeTimeout(executionBacklogQueue, 100, 100).pipe(
-            Effect.tap((items) => runInWorker(new WorkerSchema.ExecuteBulk({ items }))),
+          Queue.takeBetween(executionBacklogQueue, 1, 100).pipe(
+            Effect.tap((items) => runInWorker(new WorkerSchema.ExecuteBulk({ items: Chunk.toReadonlyArray(items) }))),
+            // NOTE we're waiting a little bit for more items to come in before executing the batch
+            Effect.tap(() => Effect.sleep(20)),
             Effect.forever,
             Effect.interruptible,
           ),
@@ -72,9 +83,9 @@ export const createStorage =
 
       const storage = {
         getInitialSnapshot: () =>
-          Effect.gen(function* ($) {
+          Effect.gen(function* () {
             // TODO replace with a proper multi-tab syncing/lock-transfer mechanism
-            yield* $(
+            yield* pipe(
               WebLock.waitForDeferredLock(lockDeferred, LIVESTORE_TAB_LOCK),
               Effect.withPerformanceMeasure('@livestore/web:waitForLock'),
             )
@@ -82,21 +93,21 @@ export const createStorage =
             // NOTE here we're trying to access the persisted data directly from the main thread which
             // ususally speeds up the init process as we don't have to wait for the worker to be ready
             // This will only work for the first tab though
-            const dataFromFile = yield* $(Effect.promise(() => getPersistedData(storageOptions, schemaHash)))
+            const dataFromFile = yield* pipe(Effect.promise(() => getPersistedData(storageOptions, schemaHash)))
             if (dataFromFile !== undefined) return dataFromFile
 
-            return yield* $(runInWorker(new WorkerSchema.Setup()))
+            return yield* pipe(runInWorker(new WorkerSchema.Setup()))
           }).pipe(Effect.withPerformanceMeasure('@livestore/web:getInitialSnapshot'), Effect.runPromise),
 
         export: () => runInWorker(new WorkerSchema.Export()).pipe(Effect.runPromise),
 
-        dangerouslyReset: () =>
-          Effect.gen(function* ($) {
-            yield* $(runInWorker(new WorkerSchema.Shutdown({})))
+        dangerouslyReset: (mode) =>
+          Effect.gen(function* () {
+            yield* pipe(runInWorker(new WorkerSchema.Shutdown({})))
 
             worker.terminate()
 
-            yield* $(Effect.promise(() => resetPersistedData(storageOptions, schemaHash)))
+            yield* pipe(Effect.promise(() => resetPersistedData(storageOptions, schemaHash, mode)))
           }).pipe(Effect.runPromise),
 
         execute: async (query, bindValues) => {
@@ -114,8 +125,8 @@ export const createStorage =
           ),
 
         shutdown: () =>
-          Effect.gen(function* ($) {
-            yield* $(
+          Effect.gen(function* () {
+            yield* pipe(
               runInWorker(new WorkerSchema.Shutdown({})).pipe(Effect.andThen(() => worker.terminate())),
               // In case the graceful shutdown didn't finish in time, we terminate the worker
               Effect.race(
@@ -126,9 +137,9 @@ export const createStorage =
               ),
             )
 
-            yield* $(Deferred.succeed(lockDeferred, undefined))
+            yield* Deferred.succeed(lockDeferred, undefined)
 
-            yield* $(Scope.close(manualScope, Exit.interrupt(FiberId.none)))
+            yield* Scope.close(manualScope, Exit.interrupt(FiberId.none))
           }).pipe(Effect.runPromise),
       } satisfies StorageDatabase
 
@@ -180,12 +191,14 @@ const getPersistedData = async (storage: WorkerSchema.StorageType, schemaHash: n
   }
 }
 
-const resetPersistedData = async (storage: WorkerSchema.StorageType, schemaHash: number) => {
+const resetPersistedData = async (storage: WorkerSchema.StorageType, schemaHash: number, resetMode: ResetMode) => {
   switch (storage.type) {
     case 'opfs': {
       const dirHandle = await getOpfsDirHandle(storage.directory)
       await dirHandle.removeEntry(getAppDbFileName(storage.filePrefix, schemaHash))
-      await dirHandle.removeEntry(getMutationlogDbFileName(storage.filePrefix))
+      if (resetMode === 'all-data') {
+        await dirHandle.removeEntry(getMutationlogDbFileName(storage.filePrefix))
+      }
       break
     }
 
@@ -196,11 +209,13 @@ const resetPersistedData = async (storage: WorkerSchema.StorageType, schemaHash:
       )
       await idbApp.deleteDb()
 
-      const idbMutationLog = new IDB(
-        storage.databaseName ?? 'livestore',
-        getMutationlogDbIdbStoreName(storage.storeNamePrefix),
-      )
-      await idbMutationLog.deleteDb()
+      if (resetMode === 'all-data') {
+        const idbMutationLog = new IDB(
+          storage.databaseName ?? 'livestore',
+          getMutationlogDbIdbStoreName(storage.storeNamePrefix),
+        )
+        await idbMutationLog.deleteDb()
+      }
 
       break
     }
@@ -209,17 +224,3 @@ const resetPersistedData = async (storage: WorkerSchema.StorageType, schemaHash:
     }
   }
 }
-
-const queueTakeTimeout = <A>(queue: Queue.Queue<A>, upToItems: number, timeout: Duration.DurationInput) =>
-  Effect.gen(function* ($) {
-    const itemsFromFirstTake = yield* $(Queue.takeUpTo(queue, upToItems))
-    if (itemsFromFirstTake.length === upToItems) return [...itemsFromFirstTake]
-
-    const missingItems = upToItems - itemsFromFirstTake.length
-
-    yield* $(Effect.sleep(timeout))
-
-    const itemsFromSecondTake = yield* $(Queue.takeUpTo(queue, missingItems))
-
-    return [...itemsFromFirstTake, ...itemsFromSecondTake]
-  })
